requirements.txt
connexion2==2.10.0
python-dotenv
connexion[swagger-ui]==2.10.0
waitress
ibm-cos-sdk
pymysql
flask